SQL DB integration & AddPurchaseOrder Stored Procedure
Switching to a real SQL database because the simulation was too
limiting.  This encompases the new database which seems to be close to
final form and working well.  AddPurchaseOrder seems to be working well
too but the frontend webpage isn't updated to work with the new database
yet.  (however Admin controller seems to be working)